Add files via upload
This KERNEL is to conduct Exploratory Data Analysis (EDA) and statistical modeling on the Titanic Dataset in order to gather insights and eventually predicting survivor(0 = Not Survived, 1 = Survived). Out of the 891 passengers that went on board the titanic, approximately 38% of them got survived where as majority 62% did not survive the disaster.